·Some bugs fixed
·Video settings dialog had a new codec option, to set the codec of the recorded
 video (default codec: XVID)

+++ threshold.cpp
@@ -1,6 +1,6 @@
 /************************************************************************************************************
 *    											                                                            *
-*    PicYou - Is a simple program to take photos. Written on C++ and using Opencv library and QT Framework *
+*    PicMoments - Is a simple program to take photos. Written on C++ and using Opencv library and QT Framework *
 *											                                                                *
 *    Copyright (C) 2017  Tiago Martins                        				                                *
 *											                                                                *
@@ -38,7 +38,7 @@ Threshold::~Threshold()
     delete ui;
 }
 
-int Threshold::getSliderVal()
+double Threshold::getSliderVal()
 {
     return ui->threshSlider->value();
 }
